Moves predict_request_handler to http/ folder
