+++ cmake/onnxruntime_hosting.cmake
@@ -33,20 +33,22 @@ file(GLOB_RECURSE onnxruntime_hosting_lib_srcs
 if(NOT WIN32)
   if(HAS_UNUSED_PARAMETER)
     set_source_files_properties(${ONNXRUNTIME_ROOT}/hosting/http/json_handling.cc PROPERTIES COMPILE_FLAGS -Wno-unused-parameter)
+    set_source_files_properties(${ONNXRUNTIME_ROOT}/hosting/http/predict_request_handler.cc PROPERTIES COMPILE_FLAGS -Wno-unused-parameter)
   endif()
 endif()
 
 file(GLOB_RECURSE onnxruntime_hosting_srcs
   "${ONNXRUNTIME_ROOT}/hosting/main.cc"
   "${ONNXRUNTIME_ROOT}/hosting/environment.cc"
-  "${ONNXRUNTIME_ROOT}/hosting/request_handler.cc"
 )
 
 # Hosting library
 add_library(onnxruntime_hosting_lib ${onnxruntime_hosting_lib_srcs})
+onnxruntime_add_include_to_target(onnxruntime_hosting_lib gsl onnx_proto hosting_proto)
 target_include_directories(onnxruntime_hosting_lib PRIVATE
   ${ONNXRUNTIME_ROOT}
   ${CMAKE_CURRENT_BINARY_DIR}/onnx
+  ${ONNXRUNTIME_ROOT}/hosting
   ${ONNXRUNTIME_ROOT}/hosting/http
   PUBLIC
   ${Boost_INCLUDE_DIR}
@@ -78,7 +80,6 @@ add_dependencies(${PROJECT_NAME} hosting_proto onnx_proto ${onnxruntime_EXTERNAL
 if(NOT WIN32)
   if(HAS_UNUSED_PARAMETER)
     set_source_files_properties("${ONNXRUNTIME_ROOT}/hosting/main.cc" PROPERTIES COMPILE_FLAGS -Wno-unused-parameter)
-    set_source_files_properties("${ONNXRUNTIME_ROOT}/hosting/request_handler.cc" PROPERTIES COMPILE_FLAGS -Wno-unused-parameter)
   endif()
 endif()
